+++ src/program.js
@@ -86,9 +86,44 @@ export class Program {
     */
     restore() {
         this.initialize();
+        this.link();
         this.checkLinkage();
 
         return this;
+    }
+
+    /**
+        Get the vertex shader source translated for the platform's API.
+
+        @method
+        @return {String} The translated vertex shader source.
+    */
+    translatedVertexSource() {
+        if (this.vertexShader) {
+            return this.vertexShader.translatedSource();
+        } else {
+            let vertexShader = new Shader(this.gl, this.appState, GL.VERTEX_SHADER, this.vertexSource);
+            let translatedSource = vertexShader.translatedSource();
+            vertexShader.delete();
+            return translatedSource;
+        }
+    }
+
+    /**
+        Get the fragment shader source translated for the platform's API.
+
+        @method
+        @return {String} The translated fragment shader source.
+    */
+    translatedFragmentSource() {
+        if (this.fragmentShader) {
+            return this.fragmentShader.translatedSource();
+        } else {
+            let fragmentShader = new Shader(this.gl, this.appState, GL.FRAGMENT_SHADER, this.fragmentSource);
+            let translatedSource = fragmentShader.translatedSource();
+            fragmentShader.delete();
+            return translatedSource;
+        }
     }
 
     /**
@@ -123,22 +158,28 @@ export class Program {
         this.samplerCount = 0;
 
         if (this.vertexSource) {
-            this.vertexShader = new Shader(this.gl, GL.VERTEX_SHADER, this.vertexSource);
+            this.vertexShader = new Shader(this.gl, this.appState, GL.VERTEX_SHADER, this.vertexSource);
         }
 
         if (this.fragmentSource) {
-            this.fragmentShader = new Shader(this.gl, GL.FRAGMENT_SHADER, this.fragmentSource);
+            this.fragmentShader = new Shader(this.gl, this.appState, GL.FRAGMENT_SHADER, this.fragmentSource);
         }
 
-        let program = this.gl.createProgram();
-        this.gl.attachShader(program, this.vertexShader.shader);
-        this.gl.attachShader(program, this.fragmentShader.shader);
+        this.program = this.gl.createProgram();
+
+        return this;
+    }
+
+    // Attach shaders and link program.
+    // Done as a separate step to avoid stalls on compileShader
+    // when doing async compile.
+    link() {
+        this.gl.attachShader(this.program, this.vertexShader.shader);
+        this.gl.attachShader(this.program, this.fragmentShader.shader);
         if (this.transformFeedbackVaryings) {
-            this.gl.transformFeedbackVaryings(program, this.transformFeedbackVaryings, GL.SEPARATE_ATTRIBS);
+            this.gl.transformFeedbackVaryings(this.program, this.transformFeedbackVaryings, GL.SEPARATE_ATTRIBS);
         }
-        this.gl.linkProgram(program);
-
-        this.program = program;
+        this.gl.linkProgram(this.program);
 
         return this;
     }
@@ -156,7 +197,7 @@ export class Program {
     // Will stall for completion.
     checkLinkage() {
         if (this.linked) {
-            return;
+            return this;
         }
 
         if (this.gl.getProgramParameter(this.program, GL.LINK_STATUS)) {
@@ -177,6 +218,8 @@ export class Program {
             this.fragmentShader.delete();
             this.fragmentShader = null;
         }
+
+        return this;
     }
 
     // Get variable handles from program