+++ src/texture.js
@@ -23,10 +23,13 @@
 
 import { 
     GL,
-    TEXTURE_FORMAT_DEFAULTS,
+    WEBGL_INFO,
+    TEXTURE_FORMATS,
     COMPRESSED_TEXTURE_TYPES,
     DUMMY_OBJECT,
-    DUMMY_UNIT_ARRAY
+    DUMMY_UNIT_ARRAY,
+    // DEPRECATED
+    TEXTURE_FORMAT_DEFAULTS
 } from "./constants.js";
 
 /**
@@ -52,7 +55,6 @@ import {
 */
 export class Texture {
     constructor(gl, appState, binding, image, width = image.width, height = image.height, depth, is3D, options = DUMMY_OBJECT) {
-        let defaultType = options.format === GL.DEPTH_COMPONENT ? GL.UNSIGNED_SHORT : GL.UNSIGNED_BYTE;
 
         this.gl = gl;
         this.binding = binding;
@@ -60,22 +62,38 @@ export class Texture {
         this.width = width || 0;
         this.height = height || 0;
         this.depth = depth || 0;
-        this.type = options.type !== undefined ? options.type : defaultType;
         this.is3D = is3D;
         this.appState = appState;
 
-        this.format = null;
-        this.internalFormat = null;
-        this.compressed = Boolean(COMPRESSED_TEXTURE_TYPES[options.format] || COMPRESSED_TEXTURE_TYPES[options.internalFormat]);
+        this.compressed = Boolean(COMPRESSED_TEXTURE_TYPES[options.internalFormat]);
         
+        if (options.format !== undefined) {
+            console.warn("Texture option 'format' is deprecated and will be removed. Use 'internalFormat' with a sized format instead.");
+            this.compressed = Boolean(COMPRESSED_TEXTURE_TYPES[options.format]);
+            if (options.type === undefined) {
+                options.type = options.format === GL.DEPTH_COMPONENT ? GL.UNSIGNED_SHORT : GL.UNSIGNED_BYTE;
+            }
+            if (options.internalFormat === undefined) {
+                if (this.compressed) {
+                    options.internalFormat = options.format;
+                } else {
+                    options.internalFormat = TEXTURE_FORMAT_DEFAULTS[options.type][options.format];
+                }
+            }
+        }
+
         if (this.compressed) {
             // For compressed textures, just need to provide one of format, internalFormat.
             // The other will be the same.
-            this.format = options.format !== undefined ? options.format : options.internalFormat;
-            this.internalFormat = options.internalFormat !== undefined ? options.internalFormat : options.format;
+            this.internalFormat = options.internalFormat;
+            this.format = this.internalFormat;
+            this.type = GL.UNSIGNED_BYTE;
         } else {
-            this.format = options.format !== undefined ? options.format : gl.RGBA;
-            this.internalFormat = options.internalFormat !== undefined ? options.internalFormat : TEXTURE_FORMAT_DEFAULTS[this.type][this.format];
+            this.internalFormat = options.internalFormat !== undefined ? options.internalFormat : GL.RGBA8;
+
+            let formatInfo = TEXTURE_FORMATS[this.internalFormat];
+            this.format = formatInfo[0];
+            this.type = options.type !== undefined ? options.type : formatInfo[1];
         }
 
         // -1 indicates unbound
@@ -94,6 +112,7 @@ export class Texture {
             maxLOD = null,
             baseLevel = null,
             maxLevel = null,
+            maxAnisotropy = 1,
             flipY = false,
             premultiplyAlpha = false
         } = options;
@@ -109,6 +128,7 @@ export class Texture {
         this.maxLOD = maxLOD;
         this.baseLevel = baseLevel;
         this.maxLevel = maxLevel;
+        this.maxAnisotropy = Math.min(maxAnisotropy, WEBGL_INFO.MAX_TEXTURE_ANISOTROPY);
         this.flipY = flipY;
         this.premultiplyAlpha = premultiplyAlpha;
         this.mipmaps = (minFilter === GL.LINEAR_MIPMAP_NEAREST || minFilter === GL.LINEAR_MIPMAP_LINEAR);
@@ -173,18 +193,25 @@ export class Texture {
         this.gl.texParameteri(this.binding, GL.TEXTURE_COMPARE_MODE, this.compareMode);
         this.gl.pixelStorei(GL.UNPACK_FLIP_Y_WEBGL, this.flipY);
         this.gl.pixelStorei(GL.UNPACK_PREMULTIPLY_ALPHA_WEBGL, this.premultiplyAlpha);
+        
         if (this.minLOD !== null) {
             this.gl.texParameterf(this.binding, GL.TEXTURE_MIN_LOD, this.minLOD);
         }
+        
         if (this.maxLOD !== null) {
             this.gl.texParameterf(this.binding, GL.TEXTURE_MAX_LOD, this.maxLOD);
         }
+        
         if (this.baseLevel !== null) {
             this.gl.texParameteri(this.binding, GL.TEXTURE_BASE_LEVEL, this.baseLevel);
         }
-
+        
         if (this.maxLevel !== null) {
             this.gl.texParameteri(this.binding, GL.TEXTURE_MAX_LEVEL, this.maxLevel);
+        }
+        
+        if (this.maxAnisotropy > 1) {
+            this.gl.texParameteri(this.binding, GL.TEXTURE_MAX_ANISOTROPY_EXT, this.maxAnisotropy);
         }
 
         let levels;